+++ src/components/data-entry/InputBase/index.ts
@@ -1,0 +1,2 @@
+export { default as InputBase } from './InputBase';
+export type { InputBaseProps } from './InputBase';
